Fix compatibility with WP 5.8
WP 5.8 requires that blocks are registered in JS now as well, the shim block for server side validation was broken as a result.

Fixes #283
